Create README.md
This application folder contains different formats of reminder codes. We are independently able to set reminders on a certain date and time. We can set reminders with blinks that come during any other activity we do. We can even remind blinks with sounds, ideally suitable for our complete attention. You can download an executable Windows application exe and execute the same to get some idea of the application for your testing.